Add runtime configuration for Prettier.
Start with some basic properties for formatting.
Explicit set wished properties, even if they are set the same as
default.
Remove the ESLinter rule for generator function stars. It lead to
trouble with the rootSaga. Cause it should not be an issue how to set
the star, its not worth the trouble here.