Define postcommit and prepush hook targets.
To avoid problems with some IDEs, define the postcommit hookt to update
git's index again.
To minimize CI test fails, require to run the test target before pushing
content to the origin peer.
This should lead to shorter/easier review processes.
Furthermore the CI tests are only for testing in a standardized
environment to avoid problems due to the developers local state.