(#29) Improve Whitelist contract addAddress function.
Allow to add the same address twice, but only (re)enable the address on
the second call and do not touch the country mappings.
In case address gets add a second time but with a different country
code, throw an exception to show the user that this is not possible.