(#16) Improve whitelist remove address functions.
The functions were really loose and don't used any requirements.
Make sure no address can be disabled that does not exist.
This is heavily important, cause else an address is added to the
whitelist mapping with 'false', but is not in the country mappings.
Make sure no empty list can be removed. This is consistent to the
addAddresses function.
Make sure not country can be removed, that has not addresses.